+++ lib/page/book_player/image_viewer.dart
@@ -5,11 +5,9 @@ import 'package:anx_reader/utils/save_img.dart';
 import 'package:anx_reader/utils/get_path/get_temp_dir.dart';
 import 'package:anx_reader/utils/log/common.dart';
 import 'package:anx_reader/utils/save_image_to_path.dart';
-import 'package:anx_reader/widgets/show_loading.dart';
+import 'package:anx_reader/utils/share_file.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
 import 'package:photo_view/photo_view.dart';
-import 'package:share_plus/share_plus.dart';
 
 class ImageViewer extends StatelessWidget {
   final String image;
@@ -80,11 +78,7 @@ class ImageViewer extends StatelessWidget {
                             "AnxReader_$bookName",
                           );
 
-                          showLoading();
-                          await SharePlus.instance.share(
-                            ShareParams(files: [XFile(path)]),
-                          );
-                          SmartDialog.dismiss();
+                          await shareFile(filePath: path);
                         },
                         icon: const Icon(Icons.share, color: Colors.white),
                       ),